Fixed: Variable reference resolution scope
Variables were being resolved as the same when contained in similar elements within a common parent.
Root problem was that in an if statement, a function was described in both the if and else statements with the same name and same parameters. The original resolution method caused all variable instances to be shared between the two functions for resolution, despite being different scopes