Fixed: Variable isReferenceTo method failed to resolve variables
Variables failed to resolve to their references to declaration. This was fixed by resolving the variable in question to it's declaration and determining if the element to check was indeed the variable it resolved to. A simple check, that proved to be very effective, and should have been done first.