Updated:Changed ObjJBlock into a interface
Needed to make the preprocessor define function into an implementation of a block interface, but it did not have an opening brace like the other elements, so I moved the common methods out (left out "getOpenBrace()"), and updated references. Code was meant to work with the code "getBlock", but to avoid name collision, the class ObjJBlock was renamed to ObjJBlockElement. This caused errors as the autogenerated method to retrieve a block was changed to getBlockElement(). This was fixed at a grammar level to make existing code work