applied temp fix to variables being erroneously flagged as undeclared
When checking to see if a variable was declared before use, it checked to see if both variable and referencing variable had the same declaration, meaning it was declared to itself. Unfortunately, this led to problems if the variable was declared inside a function, while not being assigned a value on declaration.

Both variables pointed to the same declaration -- the function declaration.

Removing check fixed problem temporarily, but a check will still need to be made